pr with logic for installing flux v1, v2 or both

+++ bootstrap/scripts/install-flux.sh
@@ -8,26 +8,32 @@ HELM_REPO=https://charts.fluxcd.io
 VALUES=deployments/fluxcd/values.yaml
 FLUX_HELM_CRD=https://raw.githubusercontent.com/fluxcd/helm-operator/chart-1.2.0/deploy/crds.yaml
 
-helm repo add fluxcd ${HELM_REPO}
-
-kubectl apply -f ${FLUX_HELM_CRD}
-kubectl -n admin delete secret flux-helm-repositories || true
-helm upgrade flux-helm-operator fluxcd/helm-operator --install --namespace admin   -f  deployments/fluxcd/helm-operator-values.yaml --version 1.2.0 --wait
-
-# Change $ENV var to correct name
-if [ $ENV = "sbox" ]; then
-ENV="sandbox"
-elif [ $ENV = "test" ]; then
-ENV="perftest"
-fi
+FLUX_V1_CLUSTER=('ithc' 'perftest' 'aat' 'demo' 'prod')
+FLUX_V2_CLUSTER=('ithc' 'perftest' 'aat' 'sbox' 'ptlsbox' 'ptl' 'preview' 'demo' 'prod')
 
 #Install kustomize
 curl -s "https://raw.githubusercontent.com/\
 kubernetes-sigs/kustomize/master/hack/install_kustomize.sh"  | bash
-TMP_DIR=/tmp/flux/${ENV}/${CLUSTER_NAME}
-mkdir -p $TMP_DIR
-# -----------------------------------------------------------
-(
+
+# ------------------------Flux V2----------------------------
+if [[ " ${FLUX_V1_CLUSTER[*]} " =~ " ${ENV} " ]]; then
+  helm repo add fluxcd ${HELM_REPO}
+
+  kubectl apply -f ${FLUX_HELM_CRD}
+  kubectl -n admin delete secret flux-helm-repositories || true
+  helm upgrade flux-helm-operator fluxcd/helm-operator --install --namespace admin   -f  deployments/fluxcd/helm-operator-values.yaml --version 1.2.0 --wait
+
+  # Change $ENV var to correct name
+  if [ $ENV = "sbox" ]; then
+  ENV="sandbox"
+  elif [ $ENV = "test" ]; then
+  ENV="perftest"
+  fi
+
+  TMP_DIR=/tmp/flux/${ENV}/${CLUSTER_NAME}
+  mkdir -p $TMP_DIR
+  # -----------------------------------------------------------
+  (
 cat <<EOF
 apiVersion: kustomize.config.k8s.io/v1beta1
 kind: Kustomization
@@ -38,52 +44,55 @@ patchesStrategicMerge:
   - https://raw.githubusercontent.com/hmcts/cnp-flux-config/master/k8s/namespaces/admin/flux/patches/${ENV}/flux.yaml
   - https://raw.githubusercontent.com/hmcts/cnp-flux-config/master/k8s/namespaces/admin/flux/patches/${ENV}/cluster-${CLUSTER_NAME}/flux.yaml
 EOF
-) > "${TMP_DIR}/kustomization.yaml"
-# -----------------------------------------------------------
+  ) > "${TMP_DIR}/kustomization.yaml"
+  # -----------------------------------------------------------
 
-kubectl apply -f https://raw.githubusercontent.com/hmcts/cnp-flux-config/master/k8s/$ENV/common/sealed-secrets/acr-credentials.yaml
+  kubectl apply -f https://raw.githubusercontent.com/hmcts/cnp-flux-config/master/k8s/$ENV/common/sealed-secrets/acr-credentials.yaml
 
-./kustomize build ${TMP_DIR} |  kubectl apply -f -
+  ./kustomize build ${TMP_DIR} |  kubectl apply -f -
+fi
 
 
 # ------------------------Flux V2----------------------------
-# FLUX_CONFIG_URL=https://raw.githubusercontent.com/hmcts/cnp-flux-config/master
+if [[ " ${FLUX_V2_CLUSTER[*]} " =~ " ${ENV} " ]]; then
+  FLUX_CONFIG_URL=https://raw.githubusercontent.com/hmcts/cnp-flux-config/master
 
-# if [ ${ENV} == "mgmt-sandbox" ]; then
-#   CLUSTER_ENV="sbox-intsvc"
-# elif [ ${ENV} == "cftptl" ]; then
-#   CLUSTER_ENV="ptl-intsvc"
-# else 
-#   CLUSTER_ENV=${ENV}
-# fi
+  if [ ${ENV} == "mgmt-sandbox" ]; then
+    CLUSTER_ENV="sbox-intsvc"
+  elif [ ${ENV} == "cftptl" ]; then
+    CLUSTER_ENV="ptl-intsvc"
+  else 
+    CLUSTER_ENV=${ENV}
+  fi
 
-# CLUSTER_NAME=$(echo ${CLUSTER_NAME} | cut -d'-' -f 2)
+  CLUSTER_NAME=$(echo ${CLUSTER_NAME} | cut -d'-' -f 2)
 
-# # Install Flux
-# kubectl apply -f ${FLUX_CONFIG_URL}/apps/flux-system/base/gotk-components.yaml
+  # Install Flux
+  kubectl apply -f ${FLUX_CONFIG_URL}/apps/flux-system/base/gotk-components.yaml
 
-# #Git credentials
-# kubectl apply -f ${FLUX_CONFIG_URL}/apps/flux-system/${CLUSTER_ENV}/base/git-credentials.yaml
+  #Git credentials
+  kubectl apply -f ${FLUX_CONFIG_URL}/apps/flux-system/${CLUSTER_ENV}/base/git-credentials.yaml
 
-# #Create Flux Sync CRDs
-# kubectl apply -f ${FLUX_CONFIG_URL}/apps/flux-system/base/flux-config-gitrepo.yaml
+  #Create Flux Sync CRDs
+  kubectl apply -f ${FLUX_CONFIG_URL}/apps/flux-system/base/flux-config-gitrepo.yaml
 
-# TMP_DIR=/tmp/flux/${CLUSTER_ENV}/${CLUSTER_NAME}
-# mkdir -p $TMP_DIR
-# # -----------------------------------------------------------
-# (
-# cat <<EOF
-# apiVersion: kustomize.config.k8s.io/v1beta1
-# kind: Kustomization
-# namespace: flux-system
-# resources:
-#     - ${FLUX_CONFIG_URL}/apps/flux-system/base/kustomize.yaml
-# patchesStrategicMerge:
-#   - ${FLUX_CONFIG_URL}/apps/flux-system/${CLUSTER_ENV}/${CLUSTER_NAME}/kustomize.yaml
-# EOF
-# ) > "${TMP_DIR}/kustomization.yaml"
-# # -----------------------------------------------------------
+  TMP_DIR=/tmp/flux/${CLUSTER_ENV}/${CLUSTER_NAME}
+  mkdir -p $TMP_DIR
+  # -----------------------------------------------------------
+  (
+cat <<EOF
+apiVersion: kustomize.config.k8s.io/v1beta1
+kind: Kustomization
+namespace: flux-system
+resources:
+  - ${FLUX_CONFIG_URL}/apps/flux-system/base/kustomize.yaml
+patchesStrategicMerge:
+  - ${FLUX_CONFIG_URL}/apps/flux-system/${CLUSTER_ENV}/${CLUSTER_NAME}/kustomize.yaml
+EOF
+  ) > "${TMP_DIR}/kustomization.yaml"
+  # -----------------------------------------------------------
 
-# ./kustomize build ${TMP_DIR} |  kubectl apply -f -
+  ./kustomize build ${TMP_DIR} |  kubectl apply -f -
+fi
 
 rm -rf kustomize
